% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_utils.R
\name{create_adjusted_outcome_rate}
\alias{create_adjusted_outcome_rate}
\title{function create outcome rate adjusted for who is vaccinated at a given time
Outcomes can either be hospitalizations or deaths, update the outcome rate
parameter to change this}
\usage{
create_adjusted_outcome_rate(
  outcome_rates,
  vaccination_schedule,
  population_sizes,
  age_susceptibility,
  contact_rates,
  min_date = "2021-02-01",
  max_date = "2021-10-01",
  hesitancy = 0.15,
  immunity_delay = 14,
  ve_outcome = 0.95,
  voc_severity = NULL
)
}
\arguments{
\item{outcome_rates}{tibble of age_groups and rate denoting either hosp or
deaths by age group}

\item{vaccination_schedule}{list of age_groups either NULL if not vaccination
in group or vector of length 2 containing string of dates in ymd format
expressing the start and stop of vaccination}

\item{population_sizes}{tibble of age_groups with numeric values denoting
population size in that age group}

\item{age_susceptibility}{tibble with columns age_group and susceptibility.
Denotes each age group susceptible to infection
thus effectively weights the contacts}

\item{contact_rates}{tibble of age_groups with numeric values denoting
contact rates in that age group}

\item{min_date}{string for min date}

\item{max_date}{string in ymd for max date}

\item{hesitancy}{numeric denoting proportion who will not take vaccine}

\item{immunity_delay}{Delay between vaccination and immunological impacts
(in days)}

\item{ve_outcome}{Vaccine efficacy for outcome}

\item{voc_severity}{risk ratio on each day describing increase in severity
due to voc. vector is 1 if no voc.}
}
\value{
list of all age groups with a vector of length min_date to max_date
where each day denotes the vaccination rate
}
\description{
function create outcome rate adjusted for who is vaccinated at a given time
Outcomes can either be hospitalizations or deaths, update the outcome rate
parameter to change this
}
