% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{create_ramp_vector}
\alias{create_ramp_vector}
\title{Create ramp vector}
\usage{
create_ramp_vector(start_date, total_days, start_ramp, ramp_length, ramp_max)
}
\arguments{
\item{start_date}{date at start of data (in ymd format)}

\item{total_days}{int Number of days in data or projection. Gives total
length of array}

\item{start_ramp}{date at start of ramp (in ymd format)}

\item{ramp_length}{length of ramp (in days)}

\item{ramp_max}{final value at end of ramp}
}
\value{
vector of length total_days
}
\description{
Creates a vector of length \code{total_days} starting at 1 and
increasing to \code{ramp_max} between \code{start_ramp} and \code{end_ramp}. Useful for
characterizing the change in transmission due to Variants of Concern.
}
\examples{
start_date <- "2020-01-01"
start_ramp <- "2020-01-05"
total_days <- 30
ramp_length <- 10
ramp_max <- 2
create_ramp_vector(
  start_date, total_days, start_ramp, ramp_length,
  ramp_max
)
}
