% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_utils.R
\name{create_segments_vector}
\alias{create_segments_vector}
\title{Create segments vector}
\usage{
create_segments_vector(start_date, total_days, segments, values = NULL)
}
\arguments{
\item{start_date}{date at start of data (in ymd format)}

\item{total_days}{int Number of days in data or projection. Gives total
length of array}

\item{segments}{vector of dates (in ymd format)}

\item{values}{vector length of segments + 1}
}
\value{
vector of length total_days
}
\description{
Create segments vector
}
\examples{
start_date <- "2020-01-01"
segments <- c("2020-01-05","2020-01-20")
total_days <- 30
values <- c(0,0.5,1)
create_segments_vector(start_date,segments,total_days,values=values)
}
