% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_utils.R
\name{create_vaccine_seqs}
\alias{create_vaccine_seqs}
\title{function create vaccine sequence based on schedule}
\usage{
create_vaccine_seqs(
  vaccination_schedule,
  population_sizes,
  min_date = "2021-02-01",
  max_date = "2021-10-01",
  hesitancy = 0.15,
  immunity_delay = 14
)
}
\arguments{
\item{vaccination_schedule}{list of age_groups. Each item in list is either:
\describe{
\item{\code{NULL}}{Group will not be vaccinated}
\item{\code{character}}{vector of two characters denoting the start and end
times in ymd format that age-group is vaccinated}
\item{\code{list}}{numeric list where each item has a \code{start_date}, an \code{end_date},
and a \code{proportion} denoting the proportion of that age-group
to vaccinate at that time. }
}}

\item{population_sizes}{list of age_groups with numeric values denoting
population size in that age group}

\item{min_date}{string for min date}

\item{max_date}{string in ymd for max date}

\item{hesitancy}{numeric denoting proportion who will not take vaccine}

\item{immunity_delay}{Delay between vaccination and immunological impacts
(in days)}
}
\value{
list of all age groups with a vector of length min_date to max_date
where each day denotes the vaccination rate
}
\description{
Function converts a vaccine schedule into a list for each age group
with the vaccination group scheduled in that age. Vaccination
schedule can either be defined for each age group as a vector representing
the start and end points of the vaccine schedule or as a list representing a
series of vaccine roll-outs for that age group with their start end date and proportion
of that population vaccinated.
}
\examples{
# example vaccination schedule that can be used.
vaccination_schedule <- list(
  "< 2" = NULL,
  "2 - 5" = NULL,
  "6 - 17" = NULL,
  "18 - 24" = c("2021-09-01", "2021-10-01"),
  "25 - 34" = c("2021-08-01", "2021-09-01"),
  "35 - 44" = c("2021-07-01", "2021-08-01"),
  "45 - 54" = c("2021-07-01", "2021-08-01"),
  "55 - 64" = c("2021-06-01", "2021-08-01"),
  "65 - 74" = c("2021-05-01", "2021-06-01"),
  "> 75" = c("2021-03-01", "2021-05-01")
)
}
