% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_seeiqr.R
\name{fit_seeiqr}
\alias{fit_seeiqr}
\title{Fit the Stan SEEIQR model}
\usage{
fit_seeiqr(
  daily_cases,
  daily_tests = NULL,
  seeiqr_model,
  obs_model = c("NB2", "Poisson", "beta-binomial"),
  forecast_days = 0,
  time_increment = 0.1,
  days_back = 45,
  R0_prior = c(log(2.6), 0.2),
  phi_prior = 1,
  f2_prior = c(0.4, 0.2),
  sampFrac2_prior = c(0.4, 0.2),
  sampFrac2_type = c("fixed", "estimated", "rw"),
  rw_sigma = 0.1,
  seed = 42,
  chains = 4,
  iter = 1000,
  sampled_fraction1 = 0.1,
  sampled_fraction2 = 0.3,
  sampled_fraction_day_change = 14,
  sampled_fraction_vec = NULL,
  fixed_f_forecast = NULL,
  day_start_fixed_f_forecast = length(daily_cases) + 1,
  pars = c(N = 5100000, D = 5, k1 = 1/5, k2 = 1, q = 0.05, r = 0.1, ur = 0.02, f1 = 1,
    start_decline = 15, end_decline = 22),
  i0 = 8,
  fsi = pars[["r"]]/(pars[["r"]] + pars[["ur"]]),
  nsi = 1 - fsi,
  state_0 = c(S = nsi * (pars[["N"]] - i0), E1 = 0.4 * nsi * i0, E2 = 0.1 * nsi * i0, I
    = 0.5 * nsi * i0, Q = 0, R = 0, Sd = fsi * (pars[["N"]] - i0), E1d = 0.4 * fsi * i0,
    E2d = 0.1 * fsi * i0, Id = 0.5 * fsi * i0, Qd = 0, Rd = 0),
  save_state_predictions = FALSE,
  delayScale = 9.85,
  delayShape = 1.73,
  ode_control = c(1e-06, 1e-05, 1e+05),
  daily_cases_omit = NULL,
  ...
)
}
\arguments{
\item{daily_cases}{A vector of daily new cases}

\item{daily_tests}{An optional vector of daily test numbers. Should include
assumed tests for the forecast. I.e. `length(daily_cases) + forecast_days =
length(daily_tests)`. Only used in the case of the beta-binomial (which
isn't working very well).}

\item{obs_model}{Type of observation model}

\item{forecast_days}{Number of days into the future to forecast. The model
will run slightly faster with fewer forecasted days.}

\item{time_increment}{Time increment for ODEs and Weibull delay-model
integration}

\item{days_back}{Number of days to go back for Weibull delay-model
integration}

\item{R0_prior}{Lognormal log mean and SD for R0 prior}

\item{phi_prior}{SD of `1/sqrt(phi) ~ Normal(0, SD)` prior, where NB2(mu,
phi) and `Var(Y) = mu + mu^2 / phi`.
<https://github.com/stan-dev/stan/wiki/Prior-Choice-Recommendations>}

\item{f2_prior}{Beta mean and SD for `f2` parameter}

\item{sampFrac2_prior}{`sampFrac` prior starting on
`sampled_fraction_day_change` if `sampFrac2_type` is "estimated" or "rw".
In the case of the random walk, this specifies the initial state prior. The
two values correspond to the mean and SD of a Beta distribution.}

\item{sampFrac2_type}{How to treat the sample fraction. Fixed, estimated, or
a constrained random walk.}

\item{rw_sigma}{The standard deviation on the sampFrac2 random walk.}

\item{seed}{MCMC seed}

\item{chains}{Number of MCMC chains}

\item{iter}{MCMC iterations per chain}

\item{sampled_fraction1}{Fraction sampled before
`sampled_fraction_day_change`}

\item{sampled_fraction2}{Fraction sampled at and after
`sampled_fraction_day_change`}

\item{sampled_fraction_day_change}{Date fraction sample changes}

\item{sampled_fraction_vec}{An optional vector of sampled fractions. Should
be of length: `length(daily_cases) + forecast_days`.}

\item{fixed_f_forecast}{Optional fixed `f` for forecast.}

\item{pars}{A named numeric vector of fixed parameter values}

\item{i0}{A scaling factor FIXME}

\item{fsi}{Fraction socially distancing. Derived parameter.}

\item{nsi}{Fraction not socially distancing. Derived parameter.}

\item{state_0}{Initial state: a named numeric vector}

\item{save_state_predictions}{Include the state predictions? `y_hat` Will
make the resulting model object much larger.}

\item{delayScale}{Weibull scale parameter for the delay in reporting.}

\item{delayShape}{Weibull shape parameter for the delay in reporting.}

\item{ode_control}{Control options for the Stan ODE solver. First is relative
difference, that absolute difference, and then maximum iterations. The
values here are the Stan defaults.}

\item{daily_cases_omit}{An optional vector of days to omit from the data
likelihood.}

\item{...}{Other arguments to pass to [rstan::sampling()].}

\item{Model}{from `rstan::stan_model(seeiqr_model)`.}
}
\value{
A named list object
}
\description{
Fit the Stan SEEIQR model
}
\author{
Sean Anderson
}
