% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{get_doubling_time}
\alias{get_doubling_time}
\title{Get the doubling time}
\usage{
get_doubling_time(
  obj,
  iter = seq_along(obj$post$R0),
  forecast_days = 25,
  window_check = 25,
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{obj}{Output from \code{\link[=fit_seir]{fit_seir()}}.}

\item{iter}{Vector of MCMC iterations to work with.}

\item{forecast_days}{Days to use in forecast.}

\item{window_check}{The window of days to use from the last day forecasted.}

\item{show_plot}{Make a diagnostic plot? Each line is a posterior sample of
the log prevalence time series. Each line should be well approximated by a
linear model}

\item{...}{Other arguments for \code{\link[=project_seir]{project_seir()}}.}
}
\value{
A vector of doubling times across iterations.
}
\description{
Get the doubling time
}
\details{
Assumes that prevalence is increasing; negative values imply halving
times. Use \code{show_plot} if you want to check the time series.
}
\examples{
# See ?project_seir
}
