% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rt.R
\name{get_rt}
\alias{get_rt}
\title{Get equivalent Rt values}
\usage{
get_rt(
  obj,
  lag = 30,
  ff = 1.25,
  forecast_days = 0,
  iter = seq_along(obj$post$R0)
)
}
\arguments{
\item{obj}{Output from \code{\link[=fit_seir]{fit_seir()}}.}

\item{lag}{Time lag for the inputs for Rt.}

\item{ff}{Unfortunately, a "fudge factor" to account for the "infectivity
kernel" not being quite right due to the fact that infectiousness is not
exponentially distributed. See J. Math. Biol. (2007) 55:803–816 Roberts and
Hesterbeek. FIXME}

\item{forecast_days}{Number of projection days.}

\item{iter}{MCMC iterations to include. Defaults to all.}
}
\value{
A data frame of state values with a column of Rt values. Number of
rows is 1 per time step. An \code{NA} for Rt where insufficient lagged data are
available.
}
\description{
Get equivalent Rt values
}
\details{
See \code{\link[=project_seir]{project_seir()}} for an example.
}
