% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{get_threshold}
\alias{get_threshold}
\title{Get the contact rate fraction threshold for increases}
\usage{
get_threshold(
  obj,
  iter = seq_along(obj$post$R0),
  forecast_days = 25,
  fs = seq(0.3, 0.8, length.out = 4),
  show_plot = TRUE,
  window_check = 25,
  ...
)
}
\arguments{
\item{obj}{Output from \code{\link[=fit_seir]{fit_seir()}}.}

\item{iter}{Vector of MCMC iterations to work with. More iterations will be
slower and not necessarily render a different result. ~200 total iterations
may be plenty.}

\item{forecast_days}{Days to use in forecast.}

\item{fs}{Contact rate fractions to test.}

\item{show_plot}{Make a diagnostic plot?}

\item{window_check}{The window of days to use from the last day forecasted.}

\item{...}{Other arguments for \code{\link[=project_seir]{project_seir()}}.}
}
\value{
The threshold value.
}
\description{
Get the contact rate fraction threshold for increases
}
\examples{
# See ?project_seir
}
