% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getlambd}
\alias{getlambd}
\title{Get expected cases}
\usage{
getlambd(out, pars, day, sampFrac, delayShape = 1.73, delayScale = 9.85)
}
\arguments{
\item{out}{Data from the ODE solver}

\item{pars}{Parameters}

\item{day}{The current date}

\item{sampFrac}{A vector of sample fractions}

\item{delayShape}{Delay shape parameter}

\item{delayScale}{Delay scale parameter}
}
\value{
The expected number of cases as a numeric value
}
\description{
Returns expected number of cases on day d
}
\author{
Caroline Colijn, Jessica Stockdale
}
