% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_projection_plot.R
\name{make_projection_plot}
\alias{make_projection_plot}
\title{Make a projection plot}
\usage{
make_projection_plot(
  models,
  cumulative = FALSE,
  first_date = "2020-03-01",
  ylim = c(0, max(out$upr) * 1.03),
  outer_quantile = c(0.05, 0.95),
  facet = TRUE,
  ncol = 1,
  cols = NULL,
  linetype = c("mu", "obs"),
  omitted_days = NULL,
  y_rep_dat = NULL,
  mu_dat = NULL,
  points_size = 1.25,
  sc_order = NULL
)
}
\arguments{
\item{models}{A list of models from [fit_seeiqr()]. If there are multiple elements than the names of the list elements will be used for the facet labels.}

\item{cumulative}{Logical for whether or not it should be a cumulative plot.}

\item{first_date}{The first date in the plot.}

\item{ylim}{The y-axis limits.}

\item{outer_quantile}{A vector representing the lower and upper outer quantiles of the credible interval.}

\item{facet}{Logical for whether or not to facet the panels. If false, then will put multiple ribbons on the same axes.}

\item{ncol}{The number of facet panel columns.}

\item{cols}{An optional vector of colours.}

\item{linetype}{Whether or not the line should represent the mean of the expectation or the median of the observation distribution.}

\item{omitted_days}{An optional vector of days to omit from the plot.}

\item{y_rep_dat}{An optional posterior predictive replicate data frame to override the input from `models`.}

\item{mu_dat}{An optional expected value data frame to override the input from `models`.}

\item{points_size}{The point size of dots.}

\item{sc_order}{An optional vector of scenario names to force the order of the panels. This character vector should match the named elements of `models`.}
}
\value{
A ggplot
}
\description{
Make a projection plot
}
