% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2prior.R
\name{post2prior}
\alias{post2prior}
\title{Extract new priors from a previous fit}
\usage{
post2prior(obj, iter = seq_len(100), time_slice = max(proj$time))
}
\arguments{
\item{obj}{A model fit with \code{\link[=fit_seir]{fit_seir()}}.}

\item{iter}{Iterations to use went creating projections to extract the state
posterior.}

\item{time_slice}{When to take the state priors at. Defaults to the
end of the fit.}
}
\description{
This function takes a model fit with \code{\link[=fit_seir]{fit_seir()}} and converts the posterior
distribution into priors that can be fed into a second model. This allows
splitting a time series into multiple blocks.
}
\examples{
# See the vignette 'Fitting case data in multiple blocks'
# for a complete example

cases <- c(
  0, 0, 1, 3, 1, 8, 0, 6, 5, 0, 7, 7, 18, 9, 22, 38, 53, 45, 40,
  77, 76, 48, 67, 78, 42, 66, 67, 92, 16, 70, 43, 53, 55, 53, 29,
  26, 37, 25, 45, 34, 40, 35
)
s1 <- c(rep(0.1, 13), rep(0.2, length(cases) - 13))
m <- fit_seir(
  cases,
  iter = 200,
  fit_type = "optimizing",
  samp_frac_fixed = s1
)
post2prior(m, iter = seq_len(10)) # just 10 for example speed
}
