% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{project_fit}
\alias{project_fit}
\title{Make projections with a fitted object}
\usage{
project_fit(obj, proj_days = 0, i = seq_len(50), f_vec = NULL)
}
\arguments{
\item{obj}{An object from \code{\link[=fit_sirbayes]{fit_sirbayes()}}.}

\item{proj_days}{The number of days to project beyond the last day of fitted
data.}

\item{i}{A vector of posterior samples to use.}

\item{f_vec}{An optional vector of \code{f} values to use after the last day of
fitted data. If left as the default of \code{NULL}, then the estimated \code{f2}
value will be used in the projection}
}
\value{
A named list with states and cases as data frames.
}
\description{
Make projections with a fitted object
}
