% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_seir.R
\name{project_seir}
\alias{project_seir}
\title{Make a projection with a SEIR fit}
\usage{
project_seir(
  obj,
  forecast_days = 30,
  f_fixed_start = NULL,
  f_fixed = NULL,
  iter = seq_along(obj$post$R0),
  return_states = FALSE,
  ...
)
}
\arguments{
\item{obj}{Output from \code{\link[=fit_seir]{fit_seir()}}.}

\item{forecast_days}{Number of projection days.}

\item{f_fixed_start}{Optional day to start changing f. Must be set if
\code{f_fixed} is set.}

\item{f_fixed}{An optional vector of fixed f values for the projection.
Should be length \code{forecast_days - (f_fixed_start - nrow(daily_cases) - 1)}. I.e. one value per day after \code{f_fixed_start} day.}

\item{iter}{MCMC iterations to include. Defaults to all.}

\item{return_states}{Logical for whether to return the ODE states.}

\item{...}{Other arguments to pass to \code{\link[rstan:sampling]{rstan::sampling()}}.}
}
\value{
A data frame:
\describe{
\item{day}{Day}
\item{data_type}{Data-type column from the case data}
\item{mu}{Expected number of cases}
\item{y_rep}{Posterior predictive replicate observation}
\item{phi}{Posterior draw of phi, the NB2 dispersion parameter, if included}
\item{.iteration}{The MCMC iteration}
}
}
\description{
Predict from a \code{\link[=fit_seir]{fit_seir()}} object, possibly with a future prediction. By default, the
projection uses the estimated f values (fraction of normal
contacts encountered for those physical distancing). The function also includes
functionality to specify a vector of fixed f values starting on a given
future date.
}
\examples{
cases <- c(
  0, 0, 1, 3, 1, 8, 0, 6, 5, 0, 7, 7, 18, 9, 22, 38, 53, 45, 40,
  77, 76, 48, 67, 78, 42, 66, 67, 92, 16, 70, 43, 53, 55, 53, 29,
  26, 37, 25, 45, 34, 40, 35
)
# Example fixed sample fractions:
s1 <- c(rep(0.1, 13), rep(0.2, length(cases) - 13))

# To use parallel processing with multiple chains:
# options(mc.cores = parallel::detectCores() / 2)

# Only using 100 iterations and 1 chain for a quick example:
m <- fit_seir(
  cases,
  iter = 100,
  chains = 1,
  samp_frac_fixed = s1
)
print(m)

p <- project_seir(m)
p

obs_dat <- data.frame(day = seq_along(cases), value = cases)

library(magrittr) # for \%>\%
tidy_seir(p) \%>\% plot_projection(obs_dat = obs_dat)

p <- project_seir(m,
  forecast_days = 100,
  f_fixed_start = 53,
  f_fixed = c(rep(0.7, 60), rep(0.2, 30)),
  iter = 1:25
)
p
tidy_seir(p) \%>\% plot_projection(obs_dat = obs_dat)
}
