% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{socdistmodel}
\alias{socdistmodel}
\title{Social distancing model}
\usage{
socdistmodel(t, state, parms, sdtiming_function)
}
\arguments{
\item{t}{time}

\item{state}{(S, E1, E2, I, Q, R, Sd, E1d, E2d, Id, Qd, Rd) S: Susceptible,
E1: Exposed but not infectious, E2: Exposed and Infectious, I: Infectious,
can be quarantined, R: Removed. The d compartments denote socially
distanced individuals.}

\item{sdtiming_function}{timing of social distancing function}

\item{pars}{(N, D, R0, k1, k2, q, r, ur, f) f: strength of social distancing,
r/(r+ur): frac of population who are distancing}
}
\value{
Time derivatives for input to ODE solver
}
\description{
The main ODEs for evaluating in R
}
\details{
SEIR-type model with time-dependent social distancing. Social distancing
reduces frequency of contact. Individuals can move between distanced and not
distanced compartments.
}
\author{
Caroline Colijn
}
