% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-plot.R
\name{tidy_seir}
\alias{tidy_seir}
\title{Generate a 'tidy' data frame of summarized projection output}
\usage{
tidy_seir(x, resample_y_rep = 10, data_type_names = NULL)
}
\arguments{
\item{x}{Output from \code{\link[=project_seir]{project_seir()}}.}

\item{resample_y_rep}{The number of times to resample from the observation distribution. Currently only set up for NB2 models.}

\item{data_type_names}{An optional named character vector to translate the data type numbers into names. E.g. \code{c("Reported cases" = 1, "Hospitalizations" = 2)}.}
}
\value{
A data frame
}
\description{
Returns a variety of quantiles and means of the posterior predictive distribution and the expected case numbers.
}
\details{
See \code{\link[=project_seir]{project_seir()}} for an example.
}
