% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine_utils.R
\name{create_adjusted_vaccination_rollout}
\alias{create_adjusted_vaccination_rollout}
\title{function create vaccine sequence}
\usage{
create_adjusted_vaccination_rollout(
  vaccination_schedule,
  population_sizes,
  age_susceptibility,
  contact_rates,
  min_date = "2021-02-01",
  max_date = "2021-10-01",
  hesitancy = 0.15,
  immunity_delay = 14,
  tb_eff = 0.8
)
}
\arguments{
\item{vaccination_schedule}{list of age_groups either NULL if not vaccination
in group or vector of length 2 containing string of dates in ymd format
expressing the start and stop of vaccination}

\item{population_sizes}{tibble of age_groups with numeric values denoting
population size in that age group}

\item{age_susceptibility}{tibble with columns age_group and susceptibility.
Denotes each age group susceptible to infection
thus effectively weights the contacts}

\item{contact_rates}{tibble of age_groups with numeric values denoting
contact rates in that age group}

\item{min_date}{string for min date}

\item{max_date}{string in ymd for max date}

\item{hesitancy}{numeric denoting proportion who will not take vaccine}

\item{immunity_delay}{Delay between vaccination and immunological impacts
(in days)}
}
\value{
list of all age groups with a vector of length min_date to max_date
where each day denotes the vaccination rate
}
\description{
function create vaccine sequence
}
